We can now specifiy what LANG environment variable we want passed in to the Process command. It now defaults to "C" if nothing is passed in.
The issue with this is, if you are writing a tool that
integrates with a UNIX utility (that's why you are using
subexec afterall!) and your user's system has a LANG
variable set that isn't what you built the system for,
then parsing output can be an issue. 

So, a user in france runs "identify" and the output is
in french instead of English. Your library needs a standard
output to parse. Hence, "C" is used as this standard output
in most UNIX utils. Usually its the same as "en_US", but it
can vary. 

Or, you can pass in a :lang variable when you call run.

Subexec.run "echo $LANG", :lang => "en_GB"

It will set the language for that ONE command. Your 
overall environment won't be effected.

Wrote tests for all of this.